refactor(Lotto): countMatchingNumber 메서드를 int 타입으로 변경

+++ src/main/java/lotto/domain/Lotto.java
@@ -16,8 +16,8 @@ public class Lotto {
         }
     }
 
-    public long countMatchingNumber(WinningNumber winningNumber) {
-        return numbers.stream()
+    public int countMatchingNumber(WinningNumber winningNumber) {
+        return (int) numbers.stream()
                 .filter(winningNumber::contains)
                 .count();
     }
